[ELF] Drop duplicate rpath entries
[Why]
We have several libraries which we are utilizing via pkg-config.
And every library provides its own '-Wl,-rpath,/opt/desy/lib'
option, which is absolutely reasonable, since every library can
be used independently. But if used together it leads to some
ridiculously long runpath's like this:
> 0x000000000000001d (RUNPATH)            Library runpath: [/opt/desy/lib:/opt/desy/lib:/opt/desy/lib:/opt/desy/lib:/opt/desy/lib]

[How]
Add and use a new helper which only appends new path entries,
if they are not already in the path. Therefor we disassemble the
existing path into its elements, append the new element, drop all
duplicates and reassemble the path.

This is of course the most inefficient way to go, but it is also
completely self contained.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>